重构heap

+++ src/common/array.rs
@@ -5,14 +5,14 @@ pub mod array {
 
     #[derive(Debug, Clone)]
     pub struct Array {
-        pub id: u64,
+        pub id: u32,
         pub len: u32,
         pub array_type: DataType,
         pub data: Vec<StackFrameValue>,
     }
 
     impl Array {
-        pub fn new(id: u64, len: u32, array_type: DataType) -> Array {
+        pub fn new(id: u32, len: u32, array_type: DataType) -> Array {
             let mut array = Array {
                 id,
                 len,
