+++ src/common/object.rs
@@ -4,14 +4,14 @@ use crate::common::value::StackFrameValue;
 
 #[derive(Debug, Clone)]
 pub struct Object {
-    pub id: u64,
+    pub id: u32,
     pub class: usize,
     pub lock: u8,
     pub field: HashMap<String, StackFrameValue>,
 }
 
 impl Object {
-    pub fn new(id: u64, class: usize) -> Object {
+    pub fn new(id: u32, class: usize) -> Object {
         Object {
             id,
             class,
