+++ src/interpreter/instructions/opcode_array.rs
@@ -138,7 +138,7 @@ pub fn multianewarray(frame: &mut StackFrame) {
                     }
                     
                     let reference = create_array(len as u32, DataType::Array { element_type: (Box::new(array_type.clone())), depth: (dimenssion) });
-                    let mut v :Vec<u64>  = Vec::new();
+                    let mut v :Vec<u32>  = Vec::new();
                     v.push(reference);
                     for i in 1 .. dimenssion{
                         let len_value: StackFrameValue = frame.op_stack.pop().unwrap();
@@ -172,7 +172,7 @@ pub fn multianewarray(frame: &mut StackFrame) {
 }
 
 
-fn create_muti_array(reference_id:u64,len: u32 ,array_type: DataType) -> u64 {
+fn create_muti_array(reference_id:u32,len: u32 ,array_type: DataType) -> u32 {
    let newarr =  create_array(len, array_type);
    let reference = get_reference(&reference_id).unwrap();
    match reference {
