+++ src/native/native_object.rs
@@ -26,7 +26,7 @@ pub fn hash_code( frame: &mut StackFrame) {
                     let class_name = get_class_name(&object.class);
                     let class_obj = get_constant_pool_class(&class_name);
                     if class_obj.is_none() {
-                        let obj_id: u64 =   java::create_class_object(&class_name);
+                        let obj_id: u32 =   java::create_class_object(&class_name);
                         put_into_class_constant_pool(class_name.clone(), obj_id);
                         frame.op_stack.push(StackFrameValue::Reference(obj_id));
                     }else {
@@ -37,7 +37,7 @@ pub fn hash_code( frame: &mut StackFrame) {
                     //info!("{:?}",array);
                     let array_class_name = get_array_class_name(array);
                     //info!("{:?}",array_class_name);
-                    let obj_id: u64 =   java::create_class_object(&array_class_name);
+                    let obj_id: u32 =   java::create_class_object(&array_class_name);
                     frame.op_stack.push(StackFrameValue::Reference(obj_id));
                 }
             }
