+++ src/utils/java.rs
@@ -1,8 +1,8 @@
 use crate::{classfile::class::Class, common::{object::Object, param::DataType, reference::Reference, value::StackFrameValue}, runtime::runtime_data_area::{self, create_array, create_object, get_or_load_class, get_reference}};
 
-pub fn create_class_object(class_name: &String) -> u64 {
+pub fn create_class_object(class_name: &String) -> u32 {
     let class0 = get_or_load_class(&String::from("java/lang/Class"));
-    let obj_id: u64 = create_object(class0.id as usize);
+    let obj_id: u32 = create_object(class0.id as usize);
     let id = create_string_object(class_name.clone());
     let referencre: &mut Reference = get_reference(&obj_id).unwrap();
     match referencre {
@@ -61,10 +61,10 @@ pub fn convert_to_rust_string(msg: StackFrameValue) ->Option<String>{
 }
 
 
-pub fn create_string_object(str_value: String) -> u64 {
+pub fn create_string_object(str_value: String) -> u32 {
     let char_array_id = {
         let chars: Vec<char> = str_value.chars().collect();
-        let char_array_id: u64 = create_array(chars.len() as u32, DataType::Array { element_type: (Box::new(DataType::Char)), depth: (1) });
+        let char_array_id: u32 = create_array(chars.len() as u32, DataType::Array { element_type: (Box::new(DataType::Char)), depth: (1) });
         let char_array_reference = get_reference(&char_array_id).unwrap();
         match char_array_reference {
             Reference::Array(array) => {
@@ -78,7 +78,7 @@ pub fn create_string_object(str_value: String) -> u64 {
     };
     let class_name = String::from("java/lang/String");
     let class: &mut Class = runtime_data_area::get_or_load_class(&class_name);
-    let obj_id: u64 = runtime_data_area::create_object(class.id);
+    let obj_id: u32 = runtime_data_area::create_object(class.id);
     let reference = runtime_data_area::get_reference(&obj_id).unwrap();
     let object: &mut Object = match reference {
         Reference::Object(obj) => obj,
